add s21_fmod() with tests

+++ include/s21_math.h
@@ -24,7 +24,7 @@ long double s21_cos(double x);
 long double s21_exp(double x);
 double s21_fabs(double x);
 double s21_floor(double x);
-long double s21_fmod(double x, double y);
+double s21_fmod(double x, double y);
 long double s21_log(double x);
 long double s21_pow(double base, double exp);
 long double s21_sin(double x);
